Adição de cenariose em Novo AgendaProject

+++ src/main/java/com/agenda/agendaProject/service/UsuarioService.java
@@ -4,6 +4,7 @@ import com.agenda.agendaProject.model.Usuario;
 import com.agenda.agendaProject.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,24 +15,37 @@ public class UsuarioService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
-    public List<Usuario> findAll() {
+    public List<Usuario> listarUsuario() {
         return usuarioRepository.findAll();
     }
 
-    public Usuario findById(Long id) {
+    public Usuario buscarPorId(int id) {
         Optional<Usuario> usuario = usuarioRepository.findById(id);
         return usuario.orElse(null);
     }
 
-    public Usuario save(Usuario usuario) {
+    public Usuario cadastrarUsuario(Usuario usuario) {
         return usuarioRepository.save(usuario);
     }
 
-    public Usuario update(Usuario usuario) {
-        return usuarioRepository.save(usuario);
+    @Transactional
+    public Usuario atualizarUsuario(int id, Usuario usuarioAtualizado) {
+        Usuario usuarioExistente = usuarioRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+
+        usuarioExistente.setUsuario(usuarioAtualizado.getUsuario());
+        usuarioExistente.setSenha(usuarioAtualizado.getSenha());
+        usuarioExistente.setTipoAcesso(usuarioAtualizado.getTipoAcesso());
+
+        return usuarioRepository.save(usuarioExistente);
     }
 
-    public void deleteById(Long id) {
-        usuarioRepository.deleteById(id);
+    @Transactional
+    public void deletarUsuario(int   id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Usuário não encontrado");
+        }
     }
 }
